refactor #74 테스트코드 리팩토링
- 서비스계층에 있던 테스트 도메인 계층으로 이동
- 불필요한 변수 제거
